requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
anytree==2.13.0
ascii_colors==0.5.2
asyncio==3.4.3
attrs==25.3.0
Authlib==1.6.3
autograd==1.7.0
backoff==2.2.1
banks==2.1.1
bcrypt==4.3.0
beartype==0.18.5
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.7
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
extra-streamlit-components==0.1.81
filelock==3.18.0
filetype==1.2.0
firecrawl-py==1.15.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.9.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
gotrue==2.12.4
graspologic==3.4.1
graspologic-native==1.2.5
greenlet==3.2.0
griffe==1.7.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
hyppo==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lightrag-hku==1.3.1
llama-cloud==0.1.18
llama-cloud-services==0.6.12
llama-index==0.12.30
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.30
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-gemini==0.4.14
llama-index-llms-google-genai==0.1.7
llama-index-llms-openai==0.3.35
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.12
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
openai==1.74.0
packaging==24.2
pandas==2.3.2
patsy==1.0.1
pillow==10.4.0
pipmaster==0.5.4
platformdirs==4.3.7
postgrest==1.1.1
POT==0.9.5
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.7.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==4.0.2
setuptools==78.1.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
st-supabase-connection==2.1.0
statsmodels==0.14.4
storage3==0.12.1
streamlit==1.49.1
streamlit-authenticator==0.2.3
StrEnum==0.4.15
striprtf==0.0.26
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.19.0
